refactor: move Pea since not used by cli
